chore: backward compatibility; readme
