+++ README.md
@@ -49,10 +49,11 @@ pip install -r requirements.txt
 ### Из консоли
 
 ```bash
+# если установлен через `pip install rosreestr2coord` или `python setup.py install`
 rosreestr2coord -c 38:06:144003:4723
 rosreestr2coord -l ./cadastral_numbers_list.txt
-# или
-python rosreestr2coord -c 38:06:144003:4723
+# запуск из директории проекта
+python -m rosreestr2coord -c 38:06:144003:4723
 ```
 
 Во время выполнения скрипта, в директории откуда был произведен запуск будут созданы файлы и папки.
@@ -65,7 +66,7 @@ python rosreestr2coord -c 38:06:144003:4723
 * -p - путь для промежуточных файлов
 * -o - путь для полученного  geojson файла
 * -e - параметр, определяющий точность аппроксимации
-* -t - тип площади: 
+* -t - тип площади:
   * 1 - Участки
   * 2 - Кварталы
   * 3 - Районы
@@ -90,7 +91,8 @@ area.get_attrs()
 ```
 
 ## Журнал
-* 21.04.2020 - ВЕРСИЯ 4.0.4
+
+* 21.04.2020 - ВЕРСИЯ 4.0.4 by [magican](https://github.com/magican)
   * небольшой рефакторинг - подготовка к type и pep тестам
   * добавлен poetry
   * добавлен Makefile
@@ -116,3 +118,7 @@ area.get_attrs()
 * 23.05.2016 - В тестовом режиме работает восстановление полигонов с отверстиями по PNG.
 * 21.05.2016 - Были внесены изменения, чтобы вернуть работу с распознаванием точек по PNG. Упала точность, пропала способность рисовать полигоны и выделять отверстия
 * 21.05.2016 - На публичных кадастровых картах заблокировали SVG и внесли ещё некоторые изменения в работу сервисов. В связи с этим перестало работать приложение
+
+## Разработка
+
+...
